autopoll + manualpoll + lazyload + configservice

+++ src/pollingmode.cpp
@@ -7,15 +7,15 @@ using namespace std;
 
 namespace configcat {
 
-shared_ptr<PollingMode> PollingMode::autoPoll(double autoPollIntervalInSeconds,
-                                                   uint32_t maxInitWaitTimeInSeconds,
-                                                   function<void()> onConfigChanged) {
+shared_ptr<PollingMode> PollingMode::autoPoll(uint32_t autoPollIntervalInSeconds,
+                                              uint32_t maxInitWaitTimeInSeconds,
+                                              function<void()> onConfigChanged) {
     return shared_ptr<AutoPollingMode>(
         new AutoPollingMode(autoPollIntervalInSeconds, maxInitWaitTimeInSeconds, onConfigChanged)
     );
 }
 
-shared_ptr<PollingMode> PollingMode::lazyLoad(double cacheRefreshIntervalInSeconds) {
+shared_ptr<PollingMode> PollingMode::lazyLoad(uint32_t cacheRefreshIntervalInSeconds) {
     return shared_ptr<LazyLoadingMode>(new LazyLoadingMode(cacheRefreshIntervalInSeconds));
 }
